Move typecheck and simulate tasks to a new plugin, tezos-client
