+++ taqueria-plugin-tezos-client/.npmignore
@@ -1,0 +1,1 @@
+.parcel-cache/*
