+++ taqueria-plugin-tezos-client/package.json
@@ -1,0 +1,51 @@
+{
+  "name": "@taqueria/plugin-tezos-client",
+  "version": "0.2.1",
+  "description": "A taqueria plugin for utilizing tezos-client",
+  "targets": {
+    "default": {
+      "source": "./index.ts",
+      "distDir": "./",
+      "context": "node",
+      "isLibrary": true
+    }
+  },
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "npx parcel build --no-cache 2>&1",
+    "pluginInfo": "npx ts-node index.ts --taqRun pluginInfo --i18n {\"foo:\"\"bar\"}"
+  },
+  "keywords": [
+    "taqueria",
+    "tezos",
+    "build",
+    "ecad",
+    "ecadlabs",
+    "plugin",
+    "tezos-client",
+    "typecheck",
+    "simulate"
+  ],
+  "engines": {
+    "node": ">=14.5"
+  },
+  "author": "ECAD Labs",
+  "license": "Apache-2.0",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/ecadlabs/taqueria.git",
+    "directory": "taqueria-plugin-tezos-client"
+  },
+  "bugs": {
+    "url": "https://github.com/ecadlabs/taqueria/issues"
+  },
+  "homepage": "https://github.com/ecadlabs/taqueria#readme",
+  "dependencies": {
+    "@taqueria/node-sdk": "^0.2.1",
+    "fast-glob": "^3.2.7"
+  },
+  "devDependencies": {
+    "parcel": "^2.0.1",
+    "typescript": "4.5.4"
+  }
+}
