+++ tests/e2e/taqueria-plugin-flextesa.spec.ts
@@ -239,11 +239,16 @@ describe("E2E Testing for taqueria flextesa plugin",  () => {
 
 });
 
-describe("E2E Testing for taqueria typechecker and simulator tasks of the flextesa plugin", () => {
+// TODO: to be moved into a different spec file when the tezos-client plugin no longer depends on the flextesa plugin.
+// If I move it now, testing won't work.
+// Starting docker in e2e/auto-test-flextesa-plugin and then e2e/auto-test-tezos-client-plugin won't work
+// because Jest parallelizes test runs across spec files it seems.
+// The volumes end up being different, and so the tests below will complain that files don't exist.
+describe("E2E Testing for taqueria typechecker and simulator tasks of the tezos-client plugin", () => {
 
     beforeAll(async () => {
         dockerName = "local"
-        await generateTestProject(taqueriaProjectPath, ["flextesa"]);
+        await generateTestProject(taqueriaProjectPath, ["tezos-client", "flextesa"]);
         await exec(`taq start sandbox ${dockerName}`, {cwd: `./${taqueriaProjectPath}`})
     })
